mobile view için navbar geliştirildi
